refactor: 서비스 타입 enum으로 변경
