+++ api/src/main/java/kernel/maidlab/api/manager/repository/ManagerServiceTypeRepository.java
@@ -9,13 +9,14 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import kernel.maidlab.api.manager.entity.ManagerServiceType;
+import kernel.maidlab.common.enums.ServiceType;
 
 @Repository
 public interface ManagerServiceTypeRepository extends JpaRepository<ManagerServiceType, Long> {
 	List<ManagerServiceType> findByManagerId(Long managerId);
 
-	@Query(value = "SELECT s.service_type FROM manager_service ms JOIN service_type s ON ms.service_id = s.id WHERE ms.manager_id = :managerId", nativeQuery = true)
-	List<String> findServiceTypeNamesByManagerId(@Param("managerId") Long managerId);
+	@Query("SELECT ms.serviceType FROM ManagerServiceType ms WHERE ms.manager.id = :managerId")
+	List<ServiceType> findServiceTypesByManagerId(@Param("managerId") Long managerId);
 
 	@Modifying
 	@Query("DELETE FROM ManagerServiceType ms WHERE ms.manager.id = :managerId")
